Refactor Icons
