+++ web/components/FloatingWriteButton.tsx
@@ -1,5 +1,5 @@
 import Link from "next/link";
-import { PlusIcon } from "./Icons";
+import { PlusIcon } from "./General/Icons";
 
 export default function FloatingWriteButton() {
   return (
