+++ web/components/General/Dropdown.tsx
@@ -1,7 +1,7 @@
 "use client";
 
 import { ReactNode, useCallback, useEffect, useRef, useState } from "react";
-import { MoreIcon } from "../Icons";
+import { MoreIcon } from "./Icons";
 
 export interface DropdownItem {
   id?: string;
